Use proper types in prestateTracer output, split build logic, and move tests to test crate.
Replace String fields in PrestateAccountState with U256 (balance), Bytes (code),
HashMap<H256, H256> (storage), and Address map keys — matching how the rest of
the codebase represents these values. Serde handles hex formatting at the
serialization boundary instead of manual format!() calls in business logic.

Split build_account_state_map (which used a use_pre: bool flag) into
find_touched_accounts, build_pre_state_map, and build_post_state_map so each
function has a single responsibility. De-duplicate the pre_map computation that
was repeated in both branches of trace_tx_prestate.

Extract the duplicated TestDatabase into a shared test/tests/levm/test_db.rs
module, update l2_hook_tests to import it, and move the prestate tracer
regression tests to test/tests/levm/prestate_tracer_tests.rs. Remove serde
round-trip tests from crates/common/tracing.rs. Drop hex and once_cell
dependencies from ethrex-vm.
