+++ crates/vm/Cargo.toml
@@ -14,7 +14,6 @@ ethrex-rlp.workspace = true
 
 derive_more = { version = "1.0.0", features = ["full"] }
 bytes.workspace = true
-hex.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
 serde.workspace = true
@@ -22,9 +21,6 @@ rayon.workspace = true
 rustc-hash.workspace = true
 
 dyn-clone = "1.0"
-
-[dev-dependencies]
-once_cell = "1.20.2"
 
 [lib]
 path = "./lib.rs"
