+++ test/tests/levm/test_db.rs
@@ -1,0 +1,74 @@
+use ethrex_common::{
+    Address, H256, U256,
+    constants::EMPTY_TRIE_HASH,
+    types::{Account, AccountState, ChainConfig, Code, CodeMetadata},
+};
+use ethrex_levm::{db::Database, errors::DatabaseError};
+use rustc_hash::FxHashMap;
+
+/// Lightweight in-memory database for VM tests.
+///
+/// Stores accounts in a `FxHashMap<Address, Account>` and implements the
+/// `Database` trait so it can back a `GeneralizedDatabase`.
+pub struct TestDatabase {
+    pub accounts: FxHashMap<Address, Account>,
+}
+
+impl TestDatabase {
+    pub fn new() -> Self {
+        Self {
+            accounts: FxHashMap::default(),
+        }
+    }
+}
+
+impl Database for TestDatabase {
+    fn get_account_state(&self, address: Address) -> Result<AccountState, DatabaseError> {
+        Ok(self
+            .accounts
+            .get(&address)
+            .map(|acc| AccountState {
+                nonce: acc.info.nonce,
+                balance: acc.info.balance,
+                storage_root: *EMPTY_TRIE_HASH,
+                code_hash: acc.info.code_hash,
+            })
+            .unwrap_or_default())
+    }
+
+    fn get_storage_value(&self, address: Address, key: H256) -> Result<U256, DatabaseError> {
+        Ok(self
+            .accounts
+            .get(&address)
+            .and_then(|acc| acc.storage.get(&key).copied())
+            .unwrap_or_default())
+    }
+
+    fn get_block_hash(&self, _block_number: u64) -> Result<H256, DatabaseError> {
+        Ok(H256::zero())
+    }
+
+    fn get_chain_config(&self) -> Result<ChainConfig, DatabaseError> {
+        Ok(ChainConfig::default())
+    }
+
+    fn get_account_code(&self, code_hash: H256) -> Result<Code, DatabaseError> {
+        for acc in self.accounts.values() {
+            if acc.info.code_hash == code_hash {
+                return Ok(acc.code.clone());
+            }
+        }
+        Ok(Code::default())
+    }
+
+    fn get_code_metadata(&self, code_hash: H256) -> Result<CodeMetadata, DatabaseError> {
+        for acc in self.accounts.values() {
+            if acc.info.code_hash == code_hash {
+                return Ok(CodeMetadata {
+                    length: acc.code.bytecode.len() as u64,
+                });
+            }
+        }
+        Ok(CodeMetadata { length: 0 })
+    }
+}
